refactor: 상품 타입, 목데이터 분리
