+++ frontend/src/mocks/data/products.json
@@ -1,0 +1,18 @@
+[
+  {
+    "id": 1,
+    "name": "꼬북칩",
+    "price": 1500,
+    "image": "https://github.com/woowacourse-teams/2023-fun-eat/assets/78616893/1f0fd418-131c-4cf8-b540-112d762b7c34",
+    "averageRating": 4.5,
+    "reviewCount": 100
+  },
+  {
+    "id": 2,
+    "name": "새우깡",
+    "price": 1000,
+    "image": "https://github.com/woowacourse-teams/2023-fun-eat/assets/78616893/1f0fd418-131c-4cf8-b540-112d762b7c34",
+    "averageRating": 4.0,
+    "reviewCount": 55
+  }
+]
