+++ frontend/src/types/product.ts
@@ -1,0 +1,8 @@
+export interface Product {
+  id: number;
+  name: string;
+  price: number;
+  image: string;
+  averageRating: number;
+  reviewCount: number;
+}
